Changed endSeq in byterange to be the last byte +1
The endSeq in ByteRange.h referred to the last byte in the ByteRange.
For a range of 10 bytes this meant the startSeq was e.q. 0, and the endSeq
9. This was a problem for a Range with only 1 byte as the startSeq and endSeq
would both have the same value. Differentiating a 1 byte Range from a pure-ack
range therefore required checking the length value as well.

The endSeq has now been changed to be the one more, e.g. for a range
of 10 bytes the startSeq 0 and the endSeq 10. The length of a range
will therefore always be endSeq - startSeq.